Refactor knowpro examples
