+++ ts/examples/chat/src/memory/knowproDataFrame.ts
@@ -12,7 +12,7 @@ import {
     ProgressBar,
     StopWatch,
 } from "interactive-app";
-import { KnowProContext } from "./knowproMemory.js";
+import { KnowproContext } from "./knowproMemory.js";
 import { Result, success } from "typechat";
 import { ensureDir, getFileName, readAllText } from "typeagent";
 import * as kp from "knowpro";
@@ -24,7 +24,7 @@ import { argDestFile, argSourceFile } from "./common.js";
 import path from "path";
 
 export async function createKnowproDataFrameCommands(
-    context: KnowProContext,
+    context: KnowproContext,
     commands: Record<string, CommandHandler>,
 ): Promise<void> {
     //commands.kpGetSchema = getSchema;
